Update InMemoryDataStore.java
In InMemoryDataStore, WriteResult was having issues that didnt want to be solved so I rewrite the class using lists to hold the inputConfig as an integer and output as a string. the process class reads the input from InMemoryInput and converts the integer to a string. GetProcessedOutput returns the out put for the tests.